# packages/api-client/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Public Atlas API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/runs": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "List Runs",
        "description": "Every run, newest first, each with its progress and cost.",
        "operationId": "runs-list_runs",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "x-database",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "`main` (the default) or `eval`: which database the request reads.",
              "title": "X-Database"
            },
            "description": "`main` (the default) or `eval`: which database the request reads."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_RunDetail_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "runs"
        ],
        "summary": "Create Run",
        "description": "Start a run: the work due for the subjects in its filter is created, held in step mode\nand queued in auto mode.",
        "operationId": "runs-create_run",
        "parameters": [
          {
            "name": "x-database",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "`main` (the default) or `eval`: which database the request reads.",
              "title": "X-Database"
            },
            "description": "`main` (the default) or `eval`: which database the request reads."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/runs/{run_id}": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "Read Run",
        "description": "The run with its assignments counted by status and result, and what it has cost.",
        "operationId": "runs-read_run",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          },
          {
            "name": "x-database",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "`main` (the default) or `eval`: which database the request reads.",
              "title": "X-Database"
            },
            "description": "`main` (the default) or `eval`: which database the request reads."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/runs/{run_id}/pause": {
      "post": {
        "tags": [
          "runs"
        ],
        "summary": "Pause Run",
        "description": "The worker starts none of the run's assignments until it is resumed; running ones finish\ntheir session.",
        "operationId": "runs-pause_run",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          },
          {
            "name": "x-database",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "`main` (the default) or `eval`: which database the request reads.",
              "title": "X-Database"
            },
            "description": "`main` (the default) or `eval`: which database the request reads."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/runs/{run_id}/resume": {
      "post": {
        "tags": [
          "runs"
        ],
        "summary": "Resume Run",
        "operationId": "runs-resume_run",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          },
          {
            "name": "x-database",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "`main` (the default) or `eval`: which database the request reads.",
              "title": "X-Database"
            },
            "description": "`main` (the default) or `eval`: which database the request reads."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/runs/{run_id}/stop": {
      "post": {
        "tags": [
          "runs"
        ],
        "summary": "Stop Run",
        "description": "Cancel everything held or queued; running assignments finish their session.",
        "operationId": "runs-stop_run",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          },
          {
            "name": "x-database",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "`main` (the default) or `eval`: which database the request reads.",
              "title": "X-Database"
            },
            "description": "`main` (the default) or `eval`: which database the request reads."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/runs/{run_id}/release": {
      "post": {
        "tags": [
          "runs"
        ],
        "summary": "Release Assignments",
        "description": "Queue held assignments: a few at a time, of one type, or the ones named.",
        "operationId": "runs-release_assignments",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          },
          {
            "name": "x-database",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "`main` (the default) or `eval`: which database the request reads.",
              "title": "X-Database"
            },
            "description": "`main` (the default) or `eval`: which database the request reads."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReleaseInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AssignmentOutput"
                  },
                  "title": "Response Runs-Release Assignments"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/assignments": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "List Assignments",
        "description": "A page of assignments by creation, oldest first unless `order` is `desc`, with their\nsubjects named.",
        "operationId": "runs-list_assignments",
        "parameters": [
          {
            "name": "run_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Run Id"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/AssignmentStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "result",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/AssignmentResult"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Result"
            }
          },
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/AssignmentType"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Type"
            }
          },
          {
            "name": "subject_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Subject Id"
            }
          },
          {
            "name": "order",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(asc|desc)$",
              "default": "asc",
              "title": "Order"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "x-database",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "`main` (the default) or `eval`: which database the request reads.",
              "title": "X-Database"
            },
            "description": "`main` (the default) or `eval`: which database the request reads."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_AssignmentOutput_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/assignments/{assignment_id}": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "Read Assignment",
        "operationId": "runs-read_assignment",
        "parameters": [
          {
            "name": "assignment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Assignment Id"
            }
          },
          {
            "name": "x-database",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "`main` (the default) or `eval`: which database the request reads.",
              "title": "X-Database"
            },
            "description": "`main` (the default) or `eval`: which database the request reads."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssignmentDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/assignments/{assignment_id}/events": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "Read Assignment Events",
        "description": "Everything the agent saw, said and did, in order: the prompt, its words, its tool calls\nand their results, and the videos when recorded, each with a short-lived link to play it.",
        "operationId": "runs-read_assignment_events",
        "parameters": [
          {
            "name": "assignment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Assignment Id"
            }
          },
          {
            "name": "x-database",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "`main` (the default) or `eval`: which database the request reads.",
              "title": "X-Database"
            },
            "description": "`main` (the default) or `eval`: which database the request reads."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventOutput"
                  },
                  "title": "Response Runs-Read Assignment Events"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/assignments/{assignment_id}/findings": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "Read Assignment Findings",
        "description": "What the assignment saved: every quote it recorded, with the entity the quote is for and\na link to the stored page.",
        "operationId": "runs-read_assignment_findings",
        "parameters": [
          {
            "name": "assignment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Assignment Id"
            }
          },
          {
            "name": "x-database",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "`main` (the default) or `eval`: which database the request reads.",
              "title": "X-Database"
            },
            "description": "`main` (the default) or `eval`: which database the request reads."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FindingOutput"
                  },
                  "title": "Response Runs-Read Assignment Findings"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AssignmentStatus": {
        "type": "string",
        "enum": [
          "held",
          "queued",
          "running",
          "finished",
          "cancelled"
        ],
        "title": "AssignmentStatus",
        "description": "The lifecycle: is it still going."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Page_RunDetail_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/RunDetail"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[RunDetail]"
      },
...
      "RunDetail": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "country_code": {
            "type": "string",
            "title": "Country Code"
          },
          "mode": {
            "$ref": "#/components/schemas/RunMode"
          },
          "status": {
            "$ref": "#/components/schemas/RunStatus"
          },
          "filter": {
            "additionalProperties": true,
            "type": "object",
            "title": "Filter"
          },
          "is_eval": {
            "type": "boolean",
            "title": "Is Eval"
          },
          "record_video": {
            "type": "boolean",
            "title": "Record Video"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "progress": {
            "$ref": "#/components/schemas/Progress"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "country_code",
          "mode",
          "status",
          "filter",
          "is_eval",
          "record_video",
          "created_at",
          "progress"
        ],
        "title": "RunDetail"
      },
...
      "RunInput": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Name"
          },
          "country_code": {
            "type": "string",
            "pattern": "^[A-Z]{2}$",
            "title": "Country Code"
          },
          "mode": {
            "$ref": "#/components/schemas/RunMode"
          },
          "filter": {
            "$ref": "#/components/schemas/RunFilter",
            "default": {
              "administrative_levels": [],
              "institution_types": [],
              "assignment_types": [],
              "subject_ids": []
            }
          },
          "record_video": {
            "type": "boolean",
            "title": "Record Video",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "name",
          "country_code",
          "mode"
        ],
        "title": "RunInput"
      },
...
    }
  }
}